requirements.txt
argparse==1.2.1
django == 2.0.2
social-auth-app-django
python-dotenv
requests
mysqlclient
django_filter